--- src/Hadrian/Haskell/Cabal/Parse.hs-boot
@@ -1,9 +1,0 @@
-module Hadrian.Haskell.Cabal.Parse where
-
-import Context.Type
-import Development.Shake
-import Hadrian.Haskell.Cabal.CabalData
-import Hadrian.Haskell.Cabal.PackageData
-
-parseCabal :: Context -> Action CabalData
-parsePackageData :: Context -> Action PackageData
