+++ src/Hadrian/Oracles/TextFile.hs
@@ -1,56 +1,31 @@
-{-# LANGUAGE TypeFamilies #-}
 -----------------------------------------------------------------------------
 -- |
 -- Module     : Hadrian.Oracles.TextFile
--- Copyright  : (c) Andrey Mokhov 2014-2017
+-- Copyright  : (c) Andrey Mokhov 2014-2018
 -- License    : MIT (see the file LICENSE)
 -- Maintainer : andrey.mokhov@gmail.com
 -- Stability  : experimental
 --
 -- Read and parse text files, tracking their contents. This oracle can be used
 -- to read configuration or package metadata files and cache the parsing.
+-- This module exports various oracle queries, whereas the corresponing Shake
+-- rules can be found in "Hadrian.Oracles.TextFile.Rules".
 -----------------------------------------------------------------------------
 module Hadrian.Oracles.TextFile (
     readTextFile, lookupValue, lookupValueOrEmpty, lookupValueOrError,
     lookupValues, lookupValuesOrEmpty, lookupValuesOrError, lookupDependencies,
-    readCabalData, unsafeReadCabalData, readPackageData, textFileOracle
+    readCabalData, unsafeReadCabalData, readPackageData
     ) where
 
-import Control.Monad
-import qualified Data.HashMap.Strict as Map
 import Data.Maybe
 import Development.Shake
-import Development.Shake.Classes
-import Development.Shake.Config
 import GHC.Stack
 
 import Context.Type
 import Hadrian.Haskell.Cabal.CabalData
 import Hadrian.Haskell.Cabal.PackageData
-import {-# SOURCE #-} Hadrian.Haskell.Cabal.Parse
-import Hadrian.Package
+import Hadrian.Oracles.TextFile.Type
 import Hadrian.Utilities
-import Stage
-
-newtype TextFile = TextFile FilePath
-    deriving (Binary, Eq, Hashable, NFData, Show, Typeable)
-type instance RuleResult TextFile = String
-
-newtype CabalFile = CabalFile Context
-    deriving (Binary, Eq, Hashable, NFData, Show, Typeable)
-type instance RuleResult CabalFile = Maybe CabalData
-
-newtype PackageDataFile = PackageDataFile Context
-    deriving (Binary, Eq, Hashable, NFData, Show, Typeable)
-type instance RuleResult PackageDataFile = Maybe PackageData
-
-newtype KeyValue = KeyValue (FilePath, String)
-    deriving (Binary, Eq, Hashable, NFData, Show, Typeable)
-type instance RuleResult KeyValue = Maybe String
-
-newtype KeyValues = KeyValues (FilePath, String)
-    deriving (Binary, Eq, Hashable, NFData, Show, Typeable)
-type instance RuleResult KeyValues = Maybe [String]
 
 -- | Read a text file, caching and tracking the result. To read and track
 -- individual lines of a text file use 'lookupValue' and its derivatives.
@@ -111,49 +86,6 @@ unsafeReadCabalData context = fromMaybe (error msg) <$> readCabalData context
 
 -- | Read and parse a @.cabal@ file recording the obtained 'PackageData',
 -- caching and tracking the result. Note that unlike 'readCabalData' this
--- function resolves all Cabal configuration flags.
+-- function resolves all Cabal configuration flags and associated conditionals.
 readPackageData :: Context -> Action (Maybe PackageData)
 readPackageData = askOracle . PackageDataFile
-
--- | This oracle reads and parses text files to answer 'readTextFile' and
--- 'lookupValue' queries, as well as their derivatives, tracking the results.
-textFileOracle :: Rules ()
-textFileOracle = do
-    text <- newCache $ \file -> do
-        need [file]
-        putLoud $ "| TextFile oracle: reading " ++ quote file ++ "..."
-        liftIO $ readFile file
-    void $ addOracle $ \(TextFile file) -> text file
-
-    kv <- newCache $ \file -> do
-        need [file]
-        putLoud $ "| KeyValue oracle: reading " ++ quote file ++ "..."
-        liftIO $ readConfigFile file
-    void $ addOracle $ \(KeyValue (file, key)) -> Map.lookup key <$> kv file
-
-    kvs <- newCache $ \file -> do
-        need [file]
-        putLoud $ "| KeyValues oracle: reading " ++ quote file ++ "..."
-        contents <- map words <$> readFileLines file
-        return $ Map.fromList [ (key, values) | (key:values) <- contents ]
-    void $ addOracle $ \(KeyValues (file, key)) -> Map.lookup key <$> kvs file
-
-    cabal <- newCache $ \(ctx@Context {..}) ->
-        case pkgCabalFile package of
-            Just file -> do
-                need [file]
-                putLoud $ "| CabalFile oracle: reading " ++ quote file
-                       ++ " (Stage: " ++ stageString stage ++ ")..."
-                Just <$> parseCabal ctx
-            Nothing -> return Nothing
-    void $ addOracle $ \(CabalFile ctx) -> cabal ctx
-
-    confCabal <- newCache $ \(ctx@Context {..}) ->
-        case pkgCabalFile package of
-            Just file -> do
-                need [file]
-                putLoud $ "| PackageDataFile oracle: reading " ++ quote file
-                       ++ " (Stage: " ++ stageString stage ++ ")..."
-                Just <$> parsePackageData ctx
-            Nothing -> return Nothing
-    void $ addOracle $ \(PackageDataFile ctx) -> confCabal ctx
